Moved the native uint to java long converter class into the main jnajack package and updated import references to it.
