+++ src/main/java/org/jaudiolibs/jnajack/JackPort.java
@@ -33,7 +33,7 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 
 import org.jaudiolibs.jnajack.lowlevel.JackLibrary;
-import org.jaudiolibs.jnajack.util.NativeToJavaTypeConverter;
+import org.jaudiolibs.jnajack.NativeToJavaTypeConverter;
 
 /**
  * Wraps a native Jack port.
