+++ src/main/java/org/jaudiolibs/jnajack/NativeToJavaTypeConverter.java
@@ -7,29 +7,29 @@
  * under the terms of the GNU Lesser General Public License as published
  * by the Free Software Foundation; either version 2.1 of the License,
  * or (at your option) any later version.
- * 
+ *
  * This code is distributed in the hope that it will be useful, but WITHOUT
  * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
  * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Lesser General Public License
  * for more details.
- * 
+ *
  * You should have received a copy of the GNU Lesser General Public License
  * along with this work; if not, see http://www.gnu.org/licenses/
- * 
- * 
+ *
+ *
  * Please visit http://neilcsmith.net if you need additional information or
  * have any questions.
  *
  */
 
-package org.jaudiolibs.jnajack.util;
+package org.jaudiolibs.jnajack;
 
-public class NativeToJavaTypeConverter {
-    public final static long nuint32ToJlong(int nativeUint32) {
+class NativeToJavaTypeConverter {
+    public final static long nuint32ToJlong(final int nativeUint32) {
         return nativeUint32 & 0xFFFFFFFFL;
     }
 
-    public final static int jlongToNuint32(long javaLong) {
+    public final static int jlongToNuint32(final long javaLong) {
         return (int)javaLong;
     }
 }
